Rework adding attach links
This reworks the code to add attach links fixing some minor problems.

First, it replaces existing attach links. This eases development since you don't
end up with lots of attach links from previously loaded versions of rob-bugson.

Second, it's more defensive about when it runs and tries not to run when it's
not a pull request page. This sets us up for fixing rob-bugson to work with
pjax.