Iterate through all links to find the one pointing to a commit
The Refined GitHub adds a class to a link, thus rob-bugson won't find it
anymore and thus can't add the "add merge comment to bug" link.

So now instead of looking for the exact HTML element we look for a link
within the div that points to a commit (a URL that contains "/commit/")
and take that.

Fixes #32