Rework attach and merge descriptions to include repo info (#48)
This reworks the attach link to include "[repoOrg/repoName]" at the
beginning and put the PR number at the end. It also changes it to a
string template.

This reports the merge description in a similar fashion.

In this way, it's easier to see which repos had what work done in them
in a single bug.