feature: run unit tests on PRs using GitHub Actions
Initially, WillKG had to manually approve the workflow to run, as I had opened the PR from a fork.

WillKG added me as a collaborator; maybe that will improve the workflow in subsequent PR updates or in future PRs when I push a branch directly to the upstream repo and open PRs from the upstream repo instead of my fork.